feature: Add `ToggleClassExtension` for changing the class names on ajax intearaction
The newly added `ToggleClassExtension` allows class changes on the interacted elements and its children when a request begins. This can be utilized to indicate an active element before the completion of an AJAX. If an error transpires, it can revert classes to their prior state. Instructions on how to use and implement this have been detailed in README.md.