feature: Spinners are preserved during redirects
If the request ends in redirect, the spinners (and button spinners) are not removed. As the options are copied to the new (redirect) request, we only remove the `spinnerInitiator` and `btnSpinnerInitiator` fields so that the spinners are not duplicated. Once the redirect is complete, the spinners will be removed as the `spinnerQueue` and `btnSpinner` have been preserved.