fix: `SuggestExtension` should only handle its own requests
The detection of suggest requests in `SuggestExtension` has been updated. Previously, all requests within the form that had the `_suggest` property were incorrectly treated as suggest requests. This caused issues such as the suggest spinner not hiding when other AJAX requests within the form were still pending. Now, the extension checks whether it should be enabled based on the UI element triggering the request. The element must match the suggest submit button for the extension to be enabled.