fix: Prevent unintended modal redraws for non-pdModal follow-up requests
Starting from version 2.6.0, the `Pd-Modal-Opened` header is always sent with the correct value, regardless of whether the request is a PdModal request. This is intentional, allowing backend components to correctly render depending on whether they’re inside a modal.

However, this behavior introduced a bug when using for example `FollowUpUrlExtension`, which lacks information about the modal context. Since the modal is open, it might be unintentionally redrawn—despite that not being the intended behavior.

To address this, the `AjaxModalExtension` has been updated to also send the `Pd-Modal-Prevent-Redraw` header when the request is not recognized as a PdModal request (as determined by `AjaxModalExtension.isPdModalRequest()`). In most cases, follow-up URLs are not meant to redraw the modal, so this effectively resolves the issue. If a modal redraw is intended, it can be achieved using a `redirect` instead.