fix: better refresh handling in `AjaxModalExtension`
Added a condition to skip `window.history.back()` if `state.cursor` equals 0. This detects if the modal has been refreshed and then opened from another link again. In that case, when closing the modal, we only want to go back in history one step, even though the other state is also pdModal state. But because it has been refreshed (therefore shown outside the modal), it would be unexpected to go back multiple steps.